Refactor: Split up Repl.hs
This commit attempts to split up Repl.hs into modules with their own
responsibilities. This will most likely make it easier to add new
commands and features to the repl.
